requirements.txt
# Scopus Search API Dependencies v3.0
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP & Networking
requests>=2.31.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Validation & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# File Upload Support
python-multipart>=0.0.6

# Database (PostgreSQL)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Caching (Redis)
redis>=5.0.0

# For Heroku deployment
gunicorn>=21.0.0
